Sync decorated pot block entity data
