--- .github/workflows/pot-visual-sync.yml
@@ -1,43 +1,0 @@
-name: Decorated pot visual sync validation
-
-on:
-  push:
-    branches: ["fix/vanilla-interaction-parity"]
-
-permissions:
-  contents: write
-
-jobs:
-  validate-and-commit:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          ref: fix/vanilla-interaction-parity
-          fetch-depth: 0
-      - uses: actions/setup-go@v5
-        with:
-          go-version-file: go.mod
-          cache: true
-      - name: Apply visual sync changes
-        run: python3 .github/scripts/finalize_pot_visual_sync.py
-      - name: Format
-        run: |
-          files=$(git diff --name-only -- '*.go')
-          if [ -n "$files" ]; then gofmt -w $files; fi
-      - name: Tidy
-        run: go mod tidy
-      - name: Test
-        run: go test ./...
-      - name: Vet
-        run: go vet ./...
-      - name: Build
-        run: go build ./...
-      - name: Commit validated visual sync
-        run: |
-          rm .github/scripts/finalize_pot_visual_sync.py .github/workflows/pot-visual-sync.yml
-          git config user.name "Elias"
-          git config user.email "86126606+el211@users.noreply.github.com"
-          git add -A
-          git commit -m "Sync decorated pot block entity data"
-          git push origin HEAD:fix/vanilla-interaction-parity
